+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 #### WIP
-- Detect some Webpack loaders when building configuration
 - Add default test glob
+
+#### 0.2.3
+- Detect some Webpack loaders when building configuration
 
 #### 0.2.2
 - Remove unused watch glob